requirements.txt
httpx
pytest
fastapi
uvicorn[standard]
sqlmodel
alembic
asyncpg
python-jose[cryptography]
psycopg2-binary
